MASSIVE OVERHAUL !
Is there anything that I didn't reorganize and redo!?
Multi-dimensional arrays for streams, rivers, and precipitations. Redid print functions for all.
Cleaned and organized all code.
Created a header file for loading the buttons.
Deleted the demo function from main.cpp - so if you want to find the original demo function from the first video that was taken and sent, it will be in versions previous to this one.
Tried to have DELTA print, the character doesn't print (tried \u03B4), the Internet says if you want to print Greek characters it has to be as a bitmap. I tried the bitmap, it looks like garbage, but it is still here in the file. But it's useless so I may end up deleting it in the future.
OK TTYL