File cleanup
Adjusted the LED strip content to be the correct number of LEDs per stream / river & correct pin assigning.
Cleaned up the files to delete the LED Ring experiment, the tube BMPs and the delta BMP as they are not used in this project. I saved them in my personal Documents folder under the Headwaters Deleted Files.