Setting up the buttons' LEDs
Added the setup for the LED rings on the buttons in the loadButtons header file. Now board needs to be powered with 12V.
As of this commit, 12V isn't working for some reason? Everything works normal on 5v but when I plug in the 12V, the LED on the Teensy board doesn't blink like it's supposed to.