LED OVERHAUL
Changed the LEDs. Each stream and river now has their "own" LED array (pointer to the main LED strip), to be able to manipulate each movement of water as its own individual array. This was suggested to do in the very beginning by Aaron but I did not do it for unknown reasons. It is much cleaner and easier, and doesn't use any blocking for loops to turn on and change the LED colors. I deleted all the other LED functions, moving them into a file called leds in my work computer Documents folder.
Next to do is animate the LEDs.