Cleaning & fixing
Changed the setting for the LEDs to be maximum brightness.
Fixed the code to color the streams and rivers as one was wrong.
Cleaned up commented out code that isn't useful for this project.
Added comments.
Deleted the setting to black of some of the LED pixels in the recursive functions to change the lights? I do not remember why they were like that, but it helped the smoothness of the water stream like animation of the lights when it was deleted.

Stream animation works better than it did, not perfect and up to them if the rest of the crew if they want to have it as a stream animation or just the set colors with no animation.

To change between the animated and the static stream colors - in main.cpp in the switch statement, change the functions (one is commented out and one is not).